+++ app/api/api_v1/routers/cclw_ingest.py
@@ -21,18 +21,18 @@ from app.api.api_v1.schemas.document import (
 )
 from app.core.auth import get_superuser_details
 from app.core.aws import get_s3_client
-from app.core.cclw_ingestion.ingest_row_cclw import (
+from app.core.ingestion.cclw.ingest_row_cclw import (
     CCLWDocumentIngestRow,
     EventIngestRow,
 )
-from app.core.cclw_ingestion.pipeline import generate_pipeline_ingest_input
+from app.core.ingestion.cclw.pipeline import generate_pipeline_ingest_input
 from app.core.ingestion.processor import (
     initialise_context,
     get_dfc_ingestor,
     get_dfc_validator,
     get_event_ingestor,
 )
-from app.core.cclw_ingestion.reader import get_file_contents, read
+from app.core.ingestion.cclw.reader import get_file_contents, read
 from app.core.ingestion.utils import (
     IngestContext,
     Result,
