+++ app/core/ingestion/cclw/family.py
@@ -2,12 +2,10 @@ from typing import Any, Optional, cast
 
 from sqlalchemy.orm import Session
 
-from app.core.unfccc_ingestion.ingest_row_unfccc import UNFCCCDocumentIngestRow
-from app.core.unfccc_ingestion.metadata import add_metadata
+from app.core.ingestion.cclw.ingest_row_cclw import CCLWDocumentIngestRow
+from app.core.ingestion.cclw.metadata import add_metadata
 from app.core.organisation import get_organisation_taxonomy
-from app.core.unfccc_ingestion.physical_document import (
-    create_physical_document_from_row,
-)
+from app.core.ingestion.cclw.physical_document import create_physical_document_from_row
 from app.core.ingestion.utils import (
     create,
     get_or_create,
@@ -28,7 +26,7 @@ from app.db.models.law_policy import (
 
 def handle_family_from_row(
     db: Session,
-    row: UNFCCCDocumentIngestRow,
+    row: CCLWDocumentIngestRow,
     org_id: int,
     result: dict[str, Any],
 ) -> Family:
@@ -52,12 +50,10 @@ def handle_family_from_row(
 
 
 def _after_create_family(
-    db: Session, row: UNFCCCDocumentIngestRow, org_id: int, result: dict[str, Any]
+    db: Session, row: CCLWDocumentIngestRow, org_id: int, result: dict[str, Any]
 ):
     def _create_family_links(family: Family):
-        # FIXME: Is this how we are handling slugs?
-        slug = f"slug_unfccc_{row.id}"
-        family_slug = Slug(name=slug, family_import_id=family.import_id)
+        family_slug = Slug(name=row.cpr_family_slug, family_import_id=family.import_id)
 
         db.add(family_slug)
         result["family_slug"] = (to_dict(family_slug),)
@@ -76,21 +72,17 @@ def _after_create_family(
 
 def _operate_on_family(
     db: Session,
-    row: UNFCCCDocumentIngestRow,
+    row: CCLWDocumentIngestRow,
     org_id: int,
     result: dict[str, Any],
 ) -> Family:
-    # FIXME: Check this:
-    category = FamilyCategory.UNFCCC
-
-    # FIXME: Check this:
-    family_summary = f"Summary for {row.id}"
+    category = FamilyCategory(row.category.upper())
 
     geography = _get_geography(db, row)
     extra = {
         "title": row.family_name,
         "geography_id": geography.id,
-        "description": family_summary,
+        "description": row.family_summary,
         "family_category": category,
     }
 
@@ -124,7 +116,7 @@ def _operate_on_family(
 
 def handle_family_document_from_row(
     db: Session,
-    row: UNFCCCDocumentIngestRow,
+    row: CCLWDocumentIngestRow,
     family: Family,
     result: dict[str, Any],
 ) -> FamilyDocument:
@@ -175,7 +167,7 @@ def handle_family_document_from_row(
         updated = {}
 
         # If source_url changed then create a new physical_document
-        if row.documents != family_document.physical_document.source_url:
+        if row.get_first_url() != family_document.physical_document.source_url:
             physical_document = create_physical_document_from_row(db, row, result)
             family_document.physical_document = physical_document
         else:
@@ -217,7 +209,7 @@ def handle_family_document_from_row(
     return family_document
 
 
-def _get_geography(db: Session, row: UNFCCCDocumentIngestRow) -> Geography:
+def _get_geography(db: Session, row: CCLWDocumentIngestRow) -> Geography:
     geography = (
         db.query(Geography).filter(Geography.value == row.geography_iso).one_or_none()
     )
@@ -230,7 +222,7 @@ def _get_geography(db: Session, row: UNFCCCDocumentIngestRow) -> Geography:
 
 def _add_family_document_slug(
     db: Session,
-    row: UNFCCCDocumentIngestRow,
+    row: CCLWDocumentIngestRow,
     family_document: FamilyDocument,
     result: dict[str, Any],
 ) -> Slug:
