+++ app/core/ingestion/cclw/pipeline.py
