+++ app/core/ingestion/unfccc/metadata.py
@@ -1,19 +1,15 @@
 from typing import Union
-
-from app.core.cclw_ingestion.ingest_row_cclw import CCLWDocumentIngestRow
-from app.db.models.law_policy.metadata import FamilyMetadata
 from sqlalchemy.orm import Session
+
+from app.core.ingestion.metadata import MetadataJson, Taxonomy, build_metadata_field
 from app.core.ingestion.utils import Result, ResultType
-from app.core.ingestion.metadata import Taxonomy, MetadataJson, build_metadata_field
+from app.core.ingestion.unfccc.ingest_row_unfccc import UNFCCCDocumentIngestRow
+from app.db.models.law_policy.metadata import FamilyMetadata
 
 
 MAP_OF_LIST_VALUES = {
-    "sector": "sectors",
-    "instrument": "instruments",
-    "framework": "frameworks",
-    "topic": "responses",
-    "hazard": "natural_hazards",
-    "keyword": "keywords",
+    "submission_type": "submission_type",
+    "author_type": "author_type",
 }
 
 
@@ -22,9 +18,9 @@ def add_metadata(
     family_import_id: str,
     taxonomy: Taxonomy,
     taxonomy_id: int,
-    row: CCLWDocumentIngestRow,
+    row: UNFCCCDocumentIngestRow,
 ) -> bool:
-    result, metadata = build_cclw_metadata(taxonomy, row)
+    result, metadata = build_unfccc_metadata(taxonomy, row)
     if result.type == ResultType.ERROR:
         return False
 
@@ -38,8 +34,8 @@ def add_metadata(
     return True
 
 
-def build_cclw_metadata(
-    taxonomy: Taxonomy, row: CCLWDocumentIngestRow
+def build_unfccc_metadata(
+    taxonomy: Taxonomy, row: UNFCCCDocumentIngestRow
 ) -> tuple[Result, MetadataJson]:
     detail_list = []
     value: dict[str, Union[str, list[str]]] = {}
