+++ tests/core/ingestion/test_event.py
@@ -2,9 +2,9 @@ from datetime import datetime, timezone
 
 from sqlalchemy.orm import Session
 
-from app.core.cclw_ingestion.event import family_event_from_row
-from app.core.cclw_ingestion.family import handle_family_from_row
-from app.core.cclw_ingestion.ingest_row_cclw import (
+from app.core.ingestion.cclw.event import family_event_from_row
+from app.core.ingestion.cclw.family import handle_family_from_row
+from app.core.ingestion.cclw.ingest_row_cclw import (
     CCLWDocumentIngestRow,
     EventIngestRow,
 )
