+++ tests/core/ingestion/test_physical_document.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
 
-from app.core.cclw_ingestion.ingest_row_cclw import CCLWDocumentIngestRow
-from app.core.cclw_ingestion.physical_document import create_physical_document_from_row
+from app.core.ingestion.cclw.ingest_row_cclw import CCLWDocumentIngestRow
+from app.core.ingestion.cclw.physical_document import create_physical_document_from_row
 from app.db.models.document import PhysicalDocument
 from app.db.models.document.physical_document import PhysicalDocumentLanguage
 from tests.core.ingestion.helpers import (
