+++ tests/core/ingestion/test_pipeline.py
@@ -5,13 +5,13 @@ from typing import Optional
 import pytest
 
 from sqlalchemy.orm import Session
-from app.core.cclw_ingestion.ingest_row_cclw import (
+from app.core.ingestion.cclw.ingest_row_cclw import (
     CCLWDocumentIngestRow,
     EventIngestRow,
 )
-from app.core.cclw_ingestion.pipeline import generate_pipeline_ingest_input
+from app.core.ingestion.cclw.pipeline import generate_pipeline_ingest_input
 from app.core.ingestion.processor import get_dfc_ingestor, get_event_ingestor
-from app.core.cclw_ingestion.reader import read
+from app.core.ingestion.cclw.reader import read
 from app.core.ingestion.utils import CCLWIngestContext
 from tests.core.ingestion.helpers import (
     FIVE_EVENT_ROWS,
